requirements.txt
hl7==0.4.1
iniconfig==2.0.0
numpy==1.24.3
packaging==24.2
pillow==11.1.0
pluggy==1.5.0
pydicom==2.3.1
pytest==7.3.1
python-dotenv==1.0.0
